Install vainfo and mesa-va-drivers
When building for docker, the OpenGL renderer context needs to be opened
with the same driver that the point clouds are generated in (e.g. NVIDIA).
Prior to this change, the OpenGL renderer driver defaulted to LLVM and thus
caused a CUDA check error when the GUI attempted to bind a buffer.

This change also adds mesa-utils, which isn't 100% necessary but is useful
for debugging future graphics issues.

Closes #95